app/requirements.txt
simplejson==3.8.2
python-status==1.0.1
Flask==0.10.1
Flask-RESTful==0.3.5
peewee==2.8.0
gunicorn==19.4.5
